+++ FitUp-BackEnd/src/main/java/tfg/dto/usuario/UsuarioRequest.java
@@ -12,6 +12,6 @@ public class UsuarioRequest {
     private String nombre;
     private String email;
     private String password;
-    private BigDecimal pesoActual;
+    private Double pesoActual;
     private Integer altura;
 }
